fixed tiny bug and starting to eat ur ass
veraciously
also adding more card attack but cant realise why is attack arrow thingy not colliding with cards lol ?
