+++ Assets/Scripts/Cards/AttackArrow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 29e9fbf4232f7c2429a140bd1f13edbb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
