+++ Assets/Scripts/Player/PlayerView.cs
@@ -67,7 +67,7 @@ public class PlayerView : NetworkBehaviour
                 cardObject.GetComponent<CardFlipper>().Flip();
             }
 
-            cardObject.GetComponent<CardData>().card.state = CardState.Hand;
+            cardObject.GetComponent<CardData>().state = CardState.Hand;
         }
         else if (cardState == CardState.Board)
         {
@@ -82,7 +82,7 @@ public class PlayerView : NetworkBehaviour
                 cardObject.GetComponent<CardFlipper>().Flip();
             }
 
-            cardObject.GetComponent<CardData>().card.state = CardState.Board;
+            cardObject.GetComponent<CardData>().state = CardState.Board;
         }
     }
